@@ -7,19 +7,19 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading.Tasks;
 
-namespace AzureFunctionSample;
+namespace InProcessAzureFunctionSample;
 
 [ExcludeFromCodeCoverage]
-public class SampleFunction
+public class InProcessFunction
 {
     private readonly IMessageContextProcessor _messageContextProcessor;
 
-    public SampleFunction(IMessageContextProcessor messageContextProcessor)
+    public InProcessFunction(IMessageContextProcessor messageContextProcessor)
     {
         _messageContextProcessor = messageContextProcessor ?? throw new ArgumentNullException(nameof(messageContextProcessor));
     }
 
-    [FunctionName("SampleFunction")]
+    [FunctionName(nameof(InProcessFunction))]
     public async Task ProcessMessageAsync(
         [ServiceBusTrigger("%ServiceBusQueueName%", Connection = "ServiceBusConnectionString")]
         ServiceBusReceivedMessage message,
