@@ -1,20 +1,20 @@
-﻿namespace Moosesoft.Azure.Webjobs.Extensions.ServiceBus;
+﻿namespace Moosesoft.Azure.Functions.Worker.Extensions.ServiceBus;
 
 /// <summary>
-/// <see cref="MessageContext"/> implementation for use with AzureFunctions In-process Model Service Bus Trigger.
+/// <see cref="MessageContext"/> implementation for use with AzureFunctions Isolated Worker Model Service Bus Trigger.
 /// </summary>
-internal class WebJobsServiceBusReceivedMessageContext : MessageContext
+internal class ServiceBusReceivedMessageContext : MessageContext
 {
     private readonly ServiceBusMessageActions _messageActions;
 
     /// <summary>
-    /// Creates a new <see cref="WebJobsServiceBusReceivedMessageContext"/>
+    /// Creates a new <see cref="ServiceBusReceivedMessageContext"/>
     /// </summary>
     /// <param name="message">Message received from Azure Service Bus.</param>
     /// <param name="messageActions">Message actions that can be used to perform operations against the received message.</param>
     /// <param name="client">Underlying Azure Service Bus client used to process received messages.</param>
     /// <exception cref="ArgumentNullException"></exception>
-    public WebJobsServiceBusReceivedMessageContext(ServiceBusReceivedMessage message, ServiceBusMessageActions messageActions, ServiceBusClient client)
+    public ServiceBusReceivedMessageContext(ServiceBusReceivedMessage message, ServiceBusMessageActions messageActions, ServiceBusClient client)
         : base(message, client)
     {
         _messageActions = messageActions;
@@ -23,7 +23,7 @@ internal class WebJobsServiceBusReceivedMessageContext : MessageContext
     /// <inheritdoc />
     public override async Task DeadLetterMessageAsync(string reason, CancellationToken cancellationToken)
     {
-        await _messageActions.DeadLetterMessageAsync(Message, reason, cancellationToken: cancellationToken)
+        await _messageActions.DeadLetterMessageAsync(Message, deadLetterReason: reason, cancellationToken: cancellationToken)
             .ConfigureAwait(false);
     }
 
