@@ -1,4 +1,4 @@
-﻿namespace Moosesoft.Azure.Webjobs.Extensions.ServiceBus;
+﻿namespace Moosesoft.Azure.Functions.Worker.Extensions.ServiceBus;
 
 /// <summary>
 /// Provides mechanism of creating instances of <see cref="MessageContext"/>.
@@ -11,12 +11,12 @@ public static class ServiceBusReceivedMessageExtensions
     /// <param name="message">Message received from Azure Service Bus.</param>
     /// <param name="messageActions">Message actions that can be used to perform operations against the received message.</param>
     /// <param name="client">Underlying Azure Service Bus client used to process received messages.</param>
-    /// <returns><see cref="MessageContext"/></returns>
+    /// <returns><see cref="ServiceBusMessageActions"/></returns>
     public static MessageContext CreateMessageContext(
         this ServiceBusReceivedMessage message,
         ServiceBusMessageActions messageActions,
         ServiceBusClient client)
     {
-        return new WebJobsServiceBusReceivedMessageContext(message, messageActions, client);
+        return new ServiceBusReceivedMessageContext(message, messageActions, client);
     }
 }
