@@ -1,8 +1,4 @@
-﻿using Moosesoft.Azure.Messaging.ServiceBus.DelayCalculatorStrategies;
-using Moosesoft.Azure.Messaging.ServiceBus.FailurePolicies;
-
-
-namespace Moosesoft.Azure.Messaging.ServiceBus.Builders;
+﻿namespace Moosesoft.Azure.Messaging.ServiceBus.Builders;
 
 /// <summary>
 /// Used to configure and build instances of <see cref="IMessageContextProcessor"/>.
